Ignored info file in examples -f command

+++ apio/examples.py
@@ -80,15 +80,16 @@ class Examples(object):
         example_files = glob.glob(join(src_path, '*'))
         for f in example_files:
             filename = basename(f)
-            if isfile(join(dest_path, filename)):
-                click.echo('Warning: ' + filename + ' file already exists')
-                key = input('Do you want to replace it? [Y/N]: ')
-                if key == 'y' or key == 'Y':
+            if filename != 'info':
+                if isfile(join(dest_path, filename)):
+                    click.echo('Warning: ' + filename + ' file already exists')
+                    key = input('Do you want to replace it? [Y/N]: ')
+                    if key == 'y' or key == 'Y':
+                        shutil.copy(f, dest_path)
+                elif isdir(join(dest_path, filename)):
+                    click.echo('Warning: ' + filename + ' is already a directory')
+                else:
                     shutil.copy(f, dest_path)
-            elif isdir(join(dest_path, filename)):
-                click.echo('Warning: ' + filename + ' is already a directory')
-            else:
-                shutil.copy(f, dest_path)
 
     def _copy_dir(self, example, src_path, dest_path):
         click.echo(' Creating ' + example + ' directory')
